+++ thoth-api/src/model/file/crud.rs
@@ -1,7 +1,6 @@
-use super::FileType;
 use super::{
-    upload_request_headers, File, FileCleanupCandidate, FilePolicy, FileUpload, FileUploadResponse,
-    NewFile, NewFileUpload,
+    upload_request_headers, ChecksumAlgorithm, File, FileCleanupCandidate, FilePolicy, FileType,
+    FileUpload, FileUploadResponse, NewFile, NewFileUpload,
 };
 use crate::db::PgPool;
 use crate::model::{
@@ -724,7 +723,7 @@ impl FileUpload {
         ctx: &C,
         work: &Work,
         cdn_url: &str,
-        cdn_checksum: &str,
+        cdn_sha256: &str,
         featured_video_dimensions: Option<(i32, i32)>,
     ) -> ThothResult<()> {
         match self.file_type {
@@ -742,7 +741,7 @@ impl FileUpload {
                     publication_id,
                     work.landing_page.clone(),
                     cdn_url,
-                    Some(cdn_checksum.to_string()),
+                    Some(cdn_sha256.to_string()),
                 )?;
             }
             FileType::AdditionalResource => {
@@ -794,7 +793,7 @@ impl FileUpload {
         publication_id: Uuid,
         landing_page: Option<String>,
         full_text_url: &str,
-        checksum: Option<String>,
+        sha256: Option<String>,
     ) -> ThothResult<()> {
         use crate::schema::location::dsl;
 
@@ -812,7 +811,8 @@ impl FileUpload {
             patch.full_text_url = Some(full_text_url.to_string());
             patch.landing_page = landing_page;
             patch.canonical = true;
-            patch.checksum = checksum;
+            patch.checksum = sha256;
+            patch.checksum_algorithm = Some(ChecksumAlgorithm::Sha256);
             if patch.canonical {
                 patch.canonical_record_complete(ctx.db())?;
             }
@@ -834,7 +834,8 @@ impl FileUpload {
                 full_text_url: Some(full_text_url.to_string()),
                 location_platform: LocationPlatform::Thoth,
                 canonical: false,
-                checksum,
+                checksum: sha256,
+                checksum_algorithm: Some(ChecksumAlgorithm::Sha256),
             };
             let created_location = Location::create(ctx.db(), &new_location)?;
             let mut patch = PatchLocation::from(created_location.clone());
@@ -850,7 +851,8 @@ impl FileUpload {
                 full_text_url: Some(full_text_url.to_string()),
                 location_platform: LocationPlatform::Thoth,
                 canonical: true,
-                checksum,
+                checksum: sha256,
+                checksum_algorithm: Some(ChecksumAlgorithm::Sha256),
             };
             new_location.canonical_record_complete(ctx.db())?;
             Location::create(ctx.db(), &new_location)?;